Create animation system
- Load textures using ResourceStore
- Create animation sprites from spritesheet textures

ResourceStore simply loads and stores textures (under a key) from image files.

Animation has a total duration and a set of frames. A frame has a
duration and a rectangle that defines which part of the spritesheet to
show for that frame. When the animation is updated, it advances to the
current frame and sets the sprite crop frame. An Animation doesn't own a
sprite, it merely operates on one (passed by reference)

AnimatedSprites own both a sprite and an animation. AnimatedSprites have a position and can be drawn.

AnimationController manages AnimatedSprites and will remove them automatically once their animation is complete.