Store and fetch shaders from ResourceStore
- Add draw functions to GroupController and MineController
- Store shader in CircleGameObject and pass to Circle when draw
- Minor cleanup

TODO: If groups are going to dynamically change shaders and their
properties then then CircelGameObject needs to provide an inderface for
that.

TODO: Do we need to include what shader a game object is using in its
network updates? If so, does ResourceStore need to store shaders with ID keys?