Reassign groups based on players' ungroup property
- Players can now ungroup by pressing ESC.
- For now players are moved to a group at the (0,0) location. This will
change to a location near their original group probably.
- Apparently pressing ESC doesn't guarantee that you are ungrouped. I
suspect this is because of the way we send reliable updates. I will
attempt to fix this in a follow up change. Currently pressing ESC once
will flash your new group due to interpolation but it is reflected
only on the server if you're lucky. Spamming ESC eventually ungroups
lol.
- Create a new TypeDef header file