Get player id from server during registration
- Now when the game loop starts will be guaranteed to have the player id
- We get the player id by continously sending tcp commands to the server
until we get a response. We need to do this because the server doesn't
immedietely have the player id when the client registers. The server
can't send the player id over the tcp socket once it's available because
it doesn't have a socket (the socket is only available when the client
sends a tcp request).