Add basic bot functionality and related command-line opts
BOTS

Currently supported strategies:
- Random
  Moves randomly in all 8 directions. Also has the option of stopping.
- NearestGreedy
  Identifies closest mine with a resource it needs, beelines to it. If
  the mines are depleted and it's no longer possible to win, switches to
  random strategy.

COMMAND-LINE OPTS

Implemented option parsing using https://github.com/jarro2783/cxxopts
version 2.10.

To run client in NearestGreedy mode invoke it with:
./ug-client -b -s 1

Random would be `-s 0` since it's just using the enum values for
BotStrategy.

Currently headless (-x) "works" in that drawing will not occur, but when
I run a headless bot and a non-headless bot, the non-headless one does
not suree the headless one, which is concerning. Leaving it in for now
as a stepping stone.