Show ratio of resources in mine
- Rather than having a set of cells that get filled with colors as you
collect resources, have as many cells as resources collected and convey
the resource amounts relative to one another. Meaning if you have one
blue and one red, half the group will be blue and the other half red.
With three blue and one read, roughly three quarters of the group will
be blue and the one quarter red. This way we don't need to do know the
"max" amount of resources a group can hold since the cell number grows
as the resource count grows. A property of displaying it this way is
that if you see a group that's just read, there is no way to know if it
has one red or a hundred - but the moment it gathers one other resource
it becomes apparent.