Improve Internet compatibility
* Send regular no-op UDP packets for NAT traversal
* Add support for using dynamic client IP
* Add support for passing server IP from client cmd line

---NAT Traversal---
We had developed on localhost so hadn't dealt with NAT traversal
previously. We had a complex bug where the client expected the server to
make first contact on the UDP channel, but the client was expecting the
server to. The server was trying, but due to the client being behind
NAT, the NAT did not know where to send the server's packets.

In this setting, the client needs to make first contact to populate the
NAT table so that server responses will be routed properly. Furthermore,
the NAT entries expire in 1-5 minutes depending on implementation. The
fix in this commit is to keep a timestamp of the last sent NAT packet
and to ensure we send one once every 30 seconds.

---Client IP---
This basically means we can have multi-client when the server is hosted
remotely (not localhost where all clients have the same IP). It also means
that we don't hard-code the client IP anymore, requiring a recompilation and
restart every time we want to run the server with a different client.

---Passing server IP from cmd line---
Default is localhost. If you use -a or -server-ip, then you can specify
and alternative server IP for the client.