Add replay tick delta threshold
- If the client is running behind the server, we don't replay we just
apply the future game state we recieive. If the client is running ahead
of the server we apply the old game state and replay back to the client
current tick to not snap back - but we restrict how far ahead the client
can replay to so that the client doesn't get too ahead of the server.
This will result in a snap back effect but it's necessary.
- The tick delta threshold can be modified in game_settings. There is also
a boolean REPLAY that indincates whether replay should be applied at
all.
- This also fixes #186. We were sometimes deleting cached local inputs
that we might need for future replays. We now only clear up to the
latest applied server tick.