Add live tick value to terminal UI
Fast number goes brr brr. Not sure how useful this is since it goes so
fast. The one use case is to compare to client tick values, but it's so
fast I'm not sure that's actually possible. Anyway looks cool and at
least it reminds you just how fast ticks are changing!