Add ability to configure server TCP port as cmd line arg
*What*
Previously you had to update the config to hit this functionality. Now
on both client and server you can change it at will.

*Why*
We get an annoying "Failed to bind" error message when restarting the
server, especially if a client is still connected, since the connection
will need to time out and is in a TIME_WAIT state until it does. Until
the connection times out the server cannot listen on the same port, this
can take > 30 seconds. My plan is pair this change with a ruby script
for running the server which will enable us to easily run the server on
a fresh port and thus iterate much more quickly.