Show TCP port server is listening on in UI
We need this value so that we can connect on that port from the client.
This is useful for rapid iteration where server restarts cause existing
ports to be unusable for a brief period of time due to that  port being
in a TIME_WAIT state. Now, we can launch a new server on a new port,
take the port used from the server UI, and then enter the matching port
on the client.