Add launch script for server
This launch script will automatically find an open TCP port for the
server to listen on and listen there. It will even open the firewall
port if you ask it to by passing --open-port. This will be useful on
ungroup.dev where connections to the game server occur over the WAN.

Add flags for not building client, server or tests