Update constants and decrese client intr tick jump
- Remove old unused constants and move some constants form code to
config
- Instead of jumping to the smallest server state when client is too far
behind, jump to the thershold. This makes the clients state "jump" less
but on the other hand it doesn't really give the client a proper boost
to catch back up. In any case I think the latter issue should be solved
in a more elegant way (for example when the client is about to fall
behind increment the tick by more than one over many frames to speed it
up)